requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
brotli==1.2.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
cobble==0.1.4
cryptography==46.0.5
cssselect2==0.9.0
cuda-bindings==12.9.4
cuda-pathfinder==1.3.4
distlib==0.4.0
distro==1.9.0
EbookLib==0.18
einops==0.8.2
et_xmlfile==2.0.0
filelock==3.24.2
filetype==1.2.0
fonttools==4.61.1
fsspec==2026.2.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
google-genai==1.63.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
identify==2.6.16
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.0.2
mammoth==1.11.0
markdown-it-py==4.0.0
markdown2==2.5.4
markdownify==1.2.2
marker-pdf==1.10.2
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.16.0
networkx==3.6.1
nodeenv==1.10.0
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
openai==1.109.1
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pdftext==0.6.3
pillow==10.4.0
platformdirs==4.9.2
pre_commit==4.5.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.0
pydantic_core==2.41.5
pydeck==0.9.1
pydyf==0.12.1
Pygments==2.19.2
pypdfium2==4.30.0
pyphen==0.17.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.37.0
regex==2024.11.6
requests==2.32.5
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
streamlit==1.54.0
streamlit-ace==0.1.1
surya-ocr==0.17.1
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.3
transformers==4.56.1
triton==3.6.0
typer==0.24.0
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
virtualenv==20.37.0
watchdog==6.0.0
wcwidth==0.6.0
weasyprint==63.1
webencodings==0.5.1
websockets==15.0.1
xlsxwriter==3.2.9
zopfli==0.4.1
